introduce ServerModule and modularize server bootstrap
