+++ sqrl-server/sqrl-server-core/src/main/java/com/datasqrl/server/modules/ServerModule.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright © 2021 DataSQRL (contact@datasqrl.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.datasqrl.server.modules;
+
+import java.util.concurrent.CompletionStage;
+
+public interface ServerModule<C extends ServerModuleContext> {
+
+  CompletionStage<Void> configure(C ctx);
+}
