+++ sqrl-server/sqrl-server-vertx-base/src/main/java/com/datasqrl/server/modules/OAuthDiscoveryModule.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright © 2021 DataSQRL (contact@datasqrl.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.datasqrl.server.modules;
+
+import com.datasqrl.server.auth.OAuthDiscoveryHandler;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+
+public class OAuthDiscoveryModule implements ServerModule<VertxServerModuleContext> {
+
+  @Override
+  public CompletionStage<Void> configure(VertxServerModuleContext ctx) {
+    var oauthDiscovery = new OAuthDiscoveryHandler(ctx.config());
+    oauthDiscovery.registerRoutes(ctx.router());
+
+    return CompletableFuture.completedFuture(null);
+  }
+}
